+++ vistock/yf_utils.py
@@ -4,9 +4,9 @@ Utility functions for working with Yahoo Finance data.
 This module contains various utility functions for retrieving and processing
 stock data using the Yahoo Finance API via the `yfinance` library.
 """
-__version__ = "3.2"
+__version__ = "3.8"
 __author__ = "York <york.jong@gmail.com>"
-__date__ = "2024/08/26 (initial version) ~ 2024/09/08 (last revision)"
+__date__ = "2024/08/26 (initial version) ~ 2024/09/09 (last revision)"
 
 __all__ = [
     'calc_weighted_metric',
@@ -87,8 +87,12 @@ def calc_weighted_metric(financials, tickers_info, metric, weight_field):
         if (weight > 0 and financial_df is not None
                        and metric in financial_df.columns):
             # Apply forward fill to fill missing metric values
-            metric_data = financial_df[metric].ffill().values
-            metric_list.append(metric_data)
+            #metric_data = financial_df[metric].ffill()
+
+            # Apply interpolate to fill missing metric values
+            metric_data = financial_df[metric].infer_objects().interpolate()
+
+            metric_list.append(metric_data.values)
             weights.append(weight)
         else:
             print("Warning: No valid metric or "
@@ -155,15 +159,40 @@ def fetch_financials(symbol, fields=None, frequency='quarterly'):
                 'annual': ticker.financials.T,
             }[frequency]
         except KeyError:
-            raise ValueError("Frequency must be 'quarterly' or 'annual'.")
+            raise ValueError("\nFrequency must be 'quarterly' or 'annual'.")
 
         financials = financials.sort_index(ascending=True)
+
+        if financials.empty:
+            print(f"\nWarning: {symbol}: Financials data is empty, "
+                  "returning NaN-filled DataFrame.")
+            if fields:
+                return pd.DataFrame({field: [np.NaN] for field in fields})
+            else:
+                return pd.DataFrame()
+
         if fields:
+            # Check for missing fields and keep only those that exist
+            missing_fields = [field for field in fields
+                              if field not in financials.columns]
+            if missing_fields:
+                print(f"\nWarning:{symbol}: Missing fields: "
+                      f"{str(missing_fields)} will be filled with NaN.")
+
+            # Ensure all requested fields are present,
+            # adding NaNs for missing fields
+            for field in fields:
+                if field not in financials.columns:
+                    financials[field] = np.NaN
+
+            # Filter financials to include only requested fields
             financials = financials[fields]
+
         return financials
+
     except Exception as e:
-        print(f"Error fetching financials for {symbol}: {e}")
-        return pd.DataFrame()
+        print(f"\nError fetching financials for {symbol}: {e}")
+        return pd.DataFrame(np.NaN, index=[0], columns=fields)
 
 
 def download_financials(symbols, fields=None, frequency='quarterly',
@@ -291,16 +320,18 @@ def download_tickers_info(symbols, fields=None, max_workers=8, progress=True):
             for key in fields:
                 try:
                     inf[key] = info[key]
-                except KeyError as e:
-                    if key in ['previousClose', 'trailingEps', 'forwardEps'
-                               'marketCap', 'sharesOutstanding']:
-                        inf[key] = np.NaN
-                    elif key in ['sector', 'industry']:
-                        inf[key] = ''
+                except KeyError:
+                    if key in ('previousClose', 'trailingEps',
+                               'revenuePerShare', 'trailingPE',
+                               'marketCap', 'sharesOutstanding'):
+                        inf[key] = np.NaN  # Default for numeric fields
+                    elif key in ['quoteType', 'sector', 'industry']:
+                        inf[key] = ''  # Default for string fields
                     else:
-                        print(f"Error fetching data for {symbol}: {e}")
+                        inf[key] = None  # Default for other data types
+                        print(f"\nError fetching data for {symbol}: {e}")
         except Exception as e:
-            print(f"Error fetching data for {symbol}: {e}")
+            print(f"\nError fetching data for {symbol}: {e}")
         return inf
 
     info_dict = {}