+++ CHANGELOG.rst
@@ -8,6 +8,18 @@ TODO
 * To support IBD EPS RS
 * To support 3-month RS (original 1-year RS)
 
+0.5.2 [2024-09-09]
+------------------
+* mansfield branch
+  * Added fileds of 'TTM EPS', 'TTM RPS', and 'TTM PE' to the generated table
+  * Added '9 Months Algo' filed
+  * Removed Rank fields of ' 1 Week Algo', ' 1 Months Algo',
+    ' 3 Months Algo', ' 6 Months Algo'
+  * Refined functions in yf_utils module
+  * Refined relative_strength_vs_benchmark
+  * Added currency label to y-axis in plot function
+  * Switched from forward fill to interpolation for missing metric values
+
 0.5.1 [2024-09-07]
 ------------------
 * Added Chinese character support to charts using mplfinance.